♻️ refactor: TableStatus component 분리
상태를 나타내는 부분의 코드를 컴포넌트로 분리
